requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0
aiohttp>=3.8.0
aiofiles>=23.0.0
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
alembic>=1.12.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
